Reduces the width of any postcode field
This is using the govuk elements form styles defined here: https://govuk-elements.herokuapp.com/form-elements/example-form-elements/

The selector will be reliable because postcode field configuration is driven from the address lookup behaviour, and so the name of the fields will always contain `postcode`. There are some old browsers that won't support the selector, but they will revert to a standard width field which is no big deal.